Added support for 10 minute videos, adjusted thresholds in VelocityCleaner.m
The new matrix application allows for longer style 10 minute video recordings of arteries. This app utilizes a longer version of the STFT to preserve quality of the frequency spectrum while cutting down on processing time for the longer videos.

Also adjusted thresholds in VelocityCleaner.m to display less cluttered velocity envelope data. This does not impact the actual output, but rather should be used as a marker for sloppy data.